Restructured code for modularization. Removed unnecessary comments and added simple structure comments
