--- src/GraphQL/Mutations/Speaker/SpeakerMutations.js
@@ -1,9 +1,0 @@
-var addspeaker = require('./Mutations/AddSpeaker').addSpeaker;
-var updatespeaker = require('./Mutations/UpdateSpeaker').updateSpeaker;
-var deletespeaker = require('./Mutations/DeleteSpeaker').removeSpeaker;
-
-module.exports = {
-  addspeaker,
-  updatespeaker,
-  deletespeaker
-}
