+++ src/GraphQL/Queries/Queries.js
@@ -1,0 +1,10 @@
+var GraphQLObjectType = require('graphql').GraphQLObjectType;
+var speakerQueries = require('./Speaker/speakerQueries');
+
+//Pools all GraphQL queries
+var queries = Object.assign(speakerQueries);
+
+exports.Queries = new GraphQLObjectType({
+    name: 'Query',
+    fields: queries
+})
