+++ src/GraphQL/Queries/Speaker/speakerQueries.js
@@ -1,0 +1,6 @@
+var getSpeakers = require('./Queries/getSpeakers').getSpeakers;
+
+//Pools all Speaker queries
+module.exports = {
+  getSpeakers
+}
